Add VideoPage class
