+++ pages/page.py
@@ -1,10 +1,10 @@
-import urlparse
+from urlparse import urljoin
 from selenium.webdriver import Remote
 
 import constants
 
 
-class Page:
+class Page(object):
     BASE_URL = constants.BASE_URL
     PATH = ''
     driver = None  # type: Remote
@@ -13,8 +13,8 @@ class Page:
         # type: (Remote) -> None
         self.driver = driver
 
-    def open(self, relative_url):
+    def open(self, relative_url=''):
         # type: (str) -> None
-        url = urlparse.urljoin(self.BASE_URL, self.PATH, relative_url)
+        url = urljoin(urljoin(self.BASE_URL, self.PATH), relative_url)
         self.driver.get(url)
         # self.driver.maximize_window()
